more

+++ Resources/UCGrapes1/toolbox/src/controls/editor/TileManager.ts
@@ -1,3 +1,4 @@
+import { InfoType } from "../../interfaces/InfoType";
 import { ActionListPopUp } from "../../ui/views/ActionListPopUp";
 import {
   DefaultAttributes,
@@ -6,8 +7,10 @@ import {
   tileWrapperDefaultAttributes,
 } from "../../utils/default-attributes";
 import { randomIdGenerator } from "../../utils/helpers";
+import { InfoSectionController } from "../InfoSectionController";
 import { CtaManager } from "../themes/CtaManager";
 import { EditorEvents } from "./EditorEvents";
+import { InfoContentMapper } from "./InfoContentMapper";
 import { NewPageButton } from "./NewPageButton";
 import { TileMapper } from "./TileMapper";
 import { TileUpdate } from "./TileUpdate";
@@ -17,15 +20,25 @@ export class TileManager {
   editor: any;
   pageId: any;
   frameId: any;
+  pageData: any;
+  page: any;
   tileUpdate: TileUpdate;
 
-  constructor(e: MouseEvent, editor: any, pageId: any, frameId: any) {
+  constructor(
+    e: MouseEvent,
+    editor: any,
+    pageId: any,
+    frameId: any,
+    pageData: any
+  ) {
     this.event = e;
     this.editor = editor;
     this.pageId = pageId;
     this.frameId = frameId;
+    this.pageData = pageData;
     this.tileUpdate = new TileUpdate(pageId);
     (globalThis as any).tileMapper = new TileMapper(this.pageId);
+    this.page = (globalThis as any).pageData;
     this.init();
   }
 
@@ -64,10 +77,13 @@ export class TileManager {
       columnComponent.append(newRowComponent, { at: index + 1 });
       const tileId = newRowComponent.find(".template-wrapper")[0]?.getId();
 
-      (globalThis as any).tileMapper.addFreshRow(
-        newRowComponent.getId() as string,
-        tileId as string
-      );
+      if (this.page?.PageType === "Information") {
+      } else {
+        (globalThis as any).tileMapper.addFreshRow(
+          newRowComponent.getId() as string,
+          tileId as string
+        );
+      }
     }
   }
 
@@ -97,10 +113,15 @@ export class TileManager {
       const index = currentTileComponent.index();
       containerRowComponent.append(newTileComponent, { at: index + 1 });
 
-      (globalThis as any).tileMapper.addTile(
-        currentTile?.parentElement?.id as string,
-        newTileComponent.getId() as string
-      );
+      if (this.page?.PageType === "Information") {
+        this.updateInfoTileRow(containerRowComponent.getId(), "add", newTileComponent.getId());
+      } else {
+        (globalThis as any).tileMapper.addTile(
+          currentTile?.parentElement?.id as string,
+          newTileComponent.getId() as string
+        );
+      }
+
       this.tileUpdate.updateTile(containerRowComponent);
     }
   }
@@ -119,13 +140,50 @@ export class TileManager {
         tileComponent.remove();
 
         this.tileUpdate.updateTile(parentComponent);
-        (globalThis as any).tileMapper.removeTile(
-          tileComponent.getId() as string,
-          parentComponent.getId() as string
-        );
+        
+        if (this.page?.PageType === "Information") {
+          this.updateInfoTileRow(parentComponent.getId(), "delete", tileComponent.getId());
+        } else {
+          (globalThis as any).tileMapper.removeTile(
+            tileComponent.getId() as string,
+            parentComponent.getId() as string
+          );
+        }
 
         this.removeEditor(tileComponent.getId() as string);
       }
+    }
+  }
+
+  private updateInfoTileRow(
+    tileRowId: any,
+    method: "add" | "delete" = "add",
+    tileId: string
+  ) {
+    const infoContentMapper = new InfoContentMapper(this.pageId);
+    const tileSection: InfoType | null =
+      infoContentMapper.getInfoContent(tileRowId);
+    if (tileSection) {
+      if (method === "add") {
+        tileSection.Tiles?.push({
+          Id: tileId,
+          Name: "Title",
+          Text: "Title",
+          Color: "#333333",
+          Align: "left",
+        });
+      } else if (method === "delete") {
+        const tile = tileSection.Tiles?.find((tile: any) => tile.Id === tileId);
+        if (tile) {
+          const index = tileSection.Tiles?.indexOf(tile);
+          if (index !== undefined && index >= 0) {
+            tileSection.Tiles?.splice(index, 1);
+          }
+        }
+      }
+
+      const infoSectionController = new InfoSectionController();
+      infoSectionController.updateInfoMapper(tileRowId, tileSection);
     }
   }
 
@@ -134,18 +192,26 @@ export class TileManager {
     if (tileIcon) {
       const templateWrapper = tileIcon.closest(".template-wrapper");
       if (templateWrapper) {
-          const tileComponent = this.editor.Components.getWrapper().find(
-            "#" + templateWrapper?.id
-          )[0];
+        const tileComponent = this.editor.Components.getWrapper().find(
+          "#" + templateWrapper?.id
+        )[0];
 
-          if (this.checkTileHasIconOrTitle(tileComponent)) {
-              (globalThis as any).tileMapper
-                .updateTile(tileComponent.getId(), "Icon", "");
-              const iconSection = tileComponent.find(".tile-icon-section")[0];
-              if (iconSection) {
-                iconSection.addStyle({ display: "none" });
-              }
+        if (this.checkTileHasIconOrTitle(tileComponent)) {
+          if (this.page?.PageType === "Information") {
+            const infoSectionController = new InfoSectionController();
+            infoSectionController.updateInfoTileAttributes(tileComponent.parent().getId(), tileComponent.getId(), 'Icon', "");
           } else {
+            (globalThis as any).tileMapper.updateTile(
+              tileComponent.getId(),
+              "Icon",
+              ""
+            );
+          }
+          const iconSection = tileComponent.find(".tile-icon-section")[0];
+          if (iconSection) {
+            iconSection.addStyle({ display: "none" });
+          }
+        } else {
           console.warn("Tile has no icon or title");
         }
       }
@@ -159,17 +225,25 @@ export class TileManager {
     if (tileTitle) {
       const templateWrapper = tileTitle.closest(".template-wrapper");
       if (templateWrapper) {
-          const tileComponent = this.editor.Components.getWrapper().find(
-            "#" + templateWrapper?.id
-          )[0];
+        const tileComponent = this.editor.Components.getWrapper().find(
+          "#" + templateWrapper?.id
+        )[0];
 
         if (this.checkTileHasIconOrTitle(tileComponent)) {
-            (globalThis as any).tileMapper
-              .updateTile(tileComponent.getId(), "Text", "");
-            const tileSection = tileComponent.find(".tile-title-section")[0];
-            if (tileSection) {
-              tileSection.addStyle({ display: "none" });
-            }
+          if (this.page?.PageType === "Information") {
+            const infoSectionController = new InfoSectionController();
+            infoSectionController.updateInfoTileAttributes(tileComponent.parent().getId(), tileComponent.getId(), 'Text', "");
+          } else {
+            (globalThis as any).tileMapper.updateTile(
+              tileComponent.getId(),
+              "Text",
+              ""
+            );
+          }
+          const tileSection = tileComponent.find(".tile-title-section")[0];
+          if (tileSection) {
+            tileSection.addStyle({ display: "none" });
+          }
         } else {
           console.warn("Tile has no icon or title");
         }
@@ -180,7 +254,21 @@ export class TileManager {
   checkTileHasIconOrTitle(component: any): boolean {
     const parentComponent = component.parent();
     if (!parentComponent) return false;
-    const tileAttributes = (globalThis as any).tileMapper.getTile(parentComponent.getId(), component.getId());
+    let tileAttributes;
+    if (this.pageData.PageType === "Information") {
+      const tileInfoSectionAttributes: InfoType = (
+        globalThis as any
+      ).infoContentMapper.getInfoContent(parentComponent.getId());
+      tileAttributes = tileInfoSectionAttributes?.Tiles?.find(
+        (tile: any) => tile.Id === component.getId()
+      );
+    } else {
+      tileAttributes = (globalThis as any).tileMapper.getTile(
+        parentComponent.getId(),
+        component.getId()
+      );
+    }
+    console.log("tileAttributes", tileAttributes);
     if (tileAttributes) {
       if (tileAttributes.Icon && tileAttributes.Text) {
         return true;
@@ -199,23 +287,27 @@ export class TileManager {
   }
 
   removeEditor(tileId: string): void {
-    const framelist = document.querySelectorAll('.mobile-frame');
+    const framelist = document.querySelectorAll(".mobile-frame");
     framelist.forEach((frame: any) => {
-      const frameHasTile = frame.querySelector(`#${tileId}`)
+      const frameHasTile = frame.querySelector(`#${tileId}`);
       if (frameHasTile) {
-        console.log(frameHasTile)
+        console.log(frameHasTile);
       }
       if (frame.id.includes(this.frameId)) {
         let nextElement = frame.nextElementSibling;
         while (nextElement) {
           const elementToRemove = nextElement;
           nextElement = nextElement.nextElementSibling;
-          if (elementToRemove) {  
-            const thumbsList = document.querySelector(".editor-thumbs-list") as HTMLElement;
-            const thumbToRemove = thumbsList.querySelector(`div[id="${elementToRemove.id}"]`);
-            if (thumbToRemove) {  
+          if (elementToRemove) {
+            const thumbsList = document.querySelector(
+              ".editor-thumbs-list"
+            ) as HTMLElement;
+            const thumbToRemove = thumbsList.querySelector(
+              `div[id="${elementToRemove.id}"]`
+            );
+            if (thumbToRemove) {
               thumbToRemove.parentElement?.parentElement?.parentElement?.remove();
-            } 
+            }
 
             elementToRemove.remove();
             new EditorEvents().activateNavigators();
@@ -243,13 +335,19 @@ export class TileManager {
               <span ${DefaultAttributes} id="ic26t" data-gjs-type="text" class="tile-icon">Title</span>
             </div>
             <div ${DefaultAttributes} id="igtdq" data-gjs-type="default" class="tile-title-section">
-              <span ${DefaultAttributes} id="is1dw" data-gjs-type="text" class="tile-close-icon top-right selected-tile-title">×</span>
+              <span ${DefaultAttributes} id="is1dw" data-gjs-type="text" class="tile-close-title top-right selected-tile-title">×</span>
               <span ${DefaultAttributes} style="display: block" id="ic26t" data-gjs-type="text" is-hidden="false" title="Title" class="tile-title">Title</span>
             </div>
         </div>
         <button ${DefaultAttributes} id="i9sxl" data-gjs-type="default" title="Delete template" class="action-button delete-button">&minus;</button>
         <button ${DefaultAttributes} id="ifvvi" data-gjs-type="default" title="Add template right" class="action-button add-button-right">+</button>
-        <button ${DefaultAttributes} id="i4ubt" data-gjs-type="default" title="Delete template" class="action-button add-button-bottom">&plus;</button>
+        ${
+          this.page?.PageType === "Information"
+            ? ``
+            : `
+          <button ${DefaultAttributes} id="i4ubt" data-gjs-type="default" title="Add template bottom" class="action-button add-button-bottom">&plus;</button>
+        `
+        }
         <svg ${DefaultAttributes} class="tile-open-menu" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 27 27">
           <g ${DefaultAttributes} id="Group_2383" data-name="Group 2383" transform="translate(-921 -417.999)">
             <g ${DefaultAttributes} id="Group_2382" data-name="Group 2382" transform="translate(921 418)">
